Made properties work
With some adjustments and converting properties to a class of its own, the json could finally read the "values" and decode it using a method i wrote back in the main method. The JSON version was returned to me. I will probably have to clean this up somehow maybe by turning that values json into a class of it's own that would potentially add on to the Minecraft Data Class.